Implement state-to-props mapping in App.js
Not sure if this is better than App.js in the previous commit. A custom
mapping allows us to expose only the slices of state that matters. We
can also get rid of the previous shouldComponentUpdate hack that was
preventing render() from trying to access 'undefined'. Downside is that
the mapping requires the structure of 'stats' to be defined.